move structs to separate files
