+++ src/protocol_version.rs
@@ -1,0 +1,6 @@
+/// SLEEP Protocol version.
+#[derive(Debug)]
+pub enum ProtocolVersion {
+  /// The version specified as per the paper released in 2017-09.
+  V0,
+}
